Implementar arquitectura y lógica central del microservicio de Pedidos
Este commit establece la arquitectura principal y la lógica de negocio para el microservicio de Pedidos, siguiendo un enfoque híbrido que combina robustez y simplicidad.

Cambios Significativos:

- **Arquitectura de Capas**: Se definen las capas de Controlador, Servicio y Repositorio, estableciendo una clara separación de responsabilidades.

- **Modelo de Datos (Entidades y DTOs)**:
  - Se implementan las entidades JPA `Pedido` e `ItemPedido` con sus relaciones (`@OneToMany`, `@ManyToOne`, `CascadeType.ALL`).
  - Se introduce un conjunto completo de DTOs (Data Transfer Objects) para desacoplar la API del modelo de datos, mejorando la seguridad y la flexibilidad en la comunicación.

- **Controlador (`PedidoControlador`)**:
  - Expone los endpoints RESTful para todas las operaciones CRUD (`GET`, `POST`, `PUT`, `DELETE`).
  - Utiliza `ResponseEntity` para un control preciso de las respuestas y códigos de estado HTTP, siguiendo las mejores prácticas.

- **Servicio (`ServicioPedido`)**:
  - Se implementa como una clase concreta que utiliza inyección de dependencias con `@Autowired` para alinearse con el estilo del curso.
  - Contiene toda la lógica de negocio, incluyendo cálculos y la gestión de estados.
  - La conversión entre DTOs y Entidades se maneja internamente a través de métodos privados, manteniendo el servicio autocontenido.

- **Repositorios y Excepciones**:
  - Se crean las interfaces `RepositorioPedido` y `RepositorioItemPedido` que extienden de `JpaRepository`.
  - Se añade una excepción personalizada (`ExcepcionPedidoNoEncontrado`) para un manejo de errores más específico.

+++ src/main/java/com/ecomarket/pedidos/models/entity/Pedido.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -39,12 +38,12 @@ public class Pedido {
 
     @NotNull(message = "El monto total no puede ser nulo.")
     @Column(nullable = false, precision = 10, scale = 2)
-    private BigDecimal montoTotal;
+    private long montoTotal;
 
     @OneToMany(mappedBy = "pedido", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY) //????
     private List<ItemPedido> itemsPedido = new ArrayList<>();
 
-    // MÉTODOS HELPER 
+    // MÉTODOS HELPER (Pedido esta en relacion a itemPedido)
     public void agregarItemPedido(ItemPedido item) {
         if (this.itemsPedido == null) { 
             this.itemsPedido = new ArrayList<>();
@@ -59,4 +58,5 @@ public class Pedido {
             item.setPedido(null);
         }
     }
+
 }
